\name{get.program.defaults}
\alias{get.program.defaults}
\title{get.program.defaults}
\description{
Get default datasets bundled with package for test runs
}
\usage{
get.program.defaults()
}
\value{
A list with \code{program.data.dir} containing path to example program directory and \code{test.data.dir} containing path to example datasets directory
}
\author{
Syed Haider
}
\examples{
x <- get.program.defaults();
}
\keyword{ defaults }
